requirements.txt
torch>=2.3
torchvision>=0.15
tqdm>=4.65
numpy>=1.20
Pillow>=9.0
opencv-python>=4.5
matplotlib>=3.0
streamlit
